move mono_wasm_load_bytes_into_heap
