+++ src/mono/wasm/runtime/src/mono/types.ts
@@ -21,10 +21,11 @@ import {
     mono_wasm_fire_debugger_agent_message
 } from './debug'
 import { StringDecoder } from './string-decoder'
-import { mono_wasm_load_bytes_into_heap } from './roots'
 import {
     mono_load_runtime_and_bcl_args, mono_wasm_load_config,
-    mono_wasm_setenv, mono_wasm_set_runtime_options, mono_wasm_load_data_archive, mono_wasm_asm_loaded
+    mono_wasm_setenv, mono_wasm_set_runtime_options,
+    mono_wasm_load_data_archive, mono_wasm_asm_loaded,
+    mono_wasm_load_bytes_into_heap
 } from './init'
 import { prevent_timer_throttling, mono_set_timeout, schedule_background_exec } from './scheduling'
 import { mono_wasm_load_icu_data, mono_wasm_get_icudt_name, GlobalizationMode } from './icu'
@@ -95,25 +96,25 @@ export interface t_ModuleExtension {
 }
 
 export type MonoConfig = {
-	assembly_root: string, // the subfolder containing managed assemblies and pdbs
-	assets: (AssetEntry | AssetEntry | SatelliteAssemblyEntry | VfsEntry | IcuData)[], // a list of assets to load along with the runtime. each asset is a dictionary-style Object with the following properties:
-	loaded_cb: Function, // a function invoked when loading has completed
-	debug_level?: number, // Either this or the next one needs to be set
-	enable_debugging?: number, // Either this or the previous one needs to be set
-	fetch_file_cb?: Request, // a function (string) invoked to fetch a given file. If no callback is provided a default implementation appropriate for the current environment will be selected (readFileSync in node, fetch elsewhere). If no default implementation is available this call will fail.
-	globalization_mode: GlobalizationMode, // configures the runtime's globalization mode
-	assembly_list?: any, // obsolete but necessary for the check
-	runtime_assets?: any, // obsolete but necessary for the check
-	runtime_asset_sources?: any, // obsolete but necessary for the check
-	diagnostic_tracing?: boolean // enables diagnostic log messages during startup
-	remote_sources?: string[], // additional search locations for assets. Sources will be checked in sequential order until the asset is found. The string "./" indicates to load from the application directory (as with the files in assembly_list), and a fully-qualified URL like "https://example.com/" indicates that asset loads can be attempted from a remote server. Sources must end with a "/".
-	environment_variables?: {
-		[i: string]: string;
-	}, // dictionary-style Object containing environment variables
-	runtime_options?: string[], // array of runtime options as strings
-	aot_profiler_options?: AOTProfilerOptions, // dictionary-style Object. If omitted, aot profiler will not be initialized.
-	coverage_profiler_options?: CoverageProfilerOptions, // dictionary-style Object. If omitted, coverage profiler will not be initialized.
-	ignore_pdb_load_errors?: boolean
+    assembly_root: string, // the subfolder containing managed assemblies and pdbs
+    assets: (AssetEntry | AssetEntry | SatelliteAssemblyEntry | VfsEntry | IcuData)[], // a list of assets to load along with the runtime. each asset is a dictionary-style Object with the following properties:
+    loaded_cb: Function, // a function invoked when loading has completed
+    debug_level?: number, // Either this or the next one needs to be set
+    enable_debugging?: number, // Either this or the previous one needs to be set
+    fetch_file_cb?: Request, // a function (string) invoked to fetch a given file. If no callback is provided a default implementation appropriate for the current environment will be selected (readFileSync in node, fetch elsewhere). If no default implementation is available this call will fail.
+    globalization_mode: GlobalizationMode, // configures the runtime's globalization mode
+    assembly_list?: any, // obsolete but necessary for the check
+    runtime_assets?: any, // obsolete but necessary for the check
+    runtime_asset_sources?: any, // obsolete but necessary for the check
+    diagnostic_tracing?: boolean // enables diagnostic log messages during startup
+    remote_sources?: string[], // additional search locations for assets. Sources will be checked in sequential order until the asset is found. The string "./" indicates to load from the application directory (as with the files in assembly_list), and a fully-qualified URL like "https://example.com/" indicates that asset loads can be attempted from a remote server. Sources must end with a "/".
+    environment_variables?: {
+        [i: string]: string;
+    }, // dictionary-style Object containing environment variables
+    runtime_options?: string[], // array of runtime options as strings
+    aot_profiler_options?: AOTProfilerOptions, // dictionary-style Object. If omitted, aot profiler will not be initialized.
+    coverage_profiler_options?: CoverageProfilerOptions, // dictionary-style Object. If omitted, coverage profiler will not be initialized.
+    ignore_pdb_load_errors?: boolean
 };
 
 export type MonoConfigError = { message: string, error: any }
